// src/components/ui/features/command-menu/core/types.ts
export type BackgroundOption =
  | 'none'
  | 'primary'
  | 'secondary'
  | 'tertiary'
  | 'quaternary'

export type BorderColorOption =
  | 'primary'
  | 'secondary'
  | 'tertiary'
  | 'quaternary'
  | 'brand'
// ...
export interface CommandItemAction extends CommandItemBase {
  type?: 'action'
  onSelect?: () => void
}

export interface CommandItemSeparator {
  type: 'separator'
  id: string
}

export type CommandItem = CommandItemAction | CommandItemSeparator

export interface CommandGroup {
  id: string
  label: string
  items: CommandItem[]
}